expression reference manager

+++ toonz/sources/toonz/columnselection.cpp
@@ -95,6 +95,8 @@ void TColumnSelection::pasteColumns() {
 
 //-----------------------------------------------------------------------------
 void TColumnSelection::deleteColumns() {
+  if (!ColumnCmd::checkExpressionReferences(m_indices)) return;
+
   ColumnCmd::deleteColumns(m_indices, false, false);
 }
